+++ CHANGELOG.md
@@ -11,6 +11,7 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 - Added accessor methods to `SamRow` enum variants
 - Added macros for facilitating floating point comparisons under the `fuzzing` feature
 - Implemented `Arbitrary` for PHMM-related types under the `fuzzing feature
+- Added `SamHmmWriter` for writing pHMMs to SAM model files
 
 ### Changed